# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/{id}": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Get user profile by id",
                "description": "Get profile of the user by user id",
                "operationId": "getUserProfileById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "userA"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "phone_number": {
                                            "type": "string",
                                            "example": "08123456789"
                                        },
                                        "user_photo": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "photo_id"
                                                },
                                                "images": {
                                                    "type": "string",
                                                    "example": "base64imagestring"
                                                },
                                                "created_by": {
                                                    "type": "string",
                                                    "example": "userA"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025/07/03 01:46:43"
                                                },
                                                "updated_by": {
                                                    "type": "string",
                                                    "example": "userA"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025/07/03 01:46:43"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-photo": {
            "post": {
                "tags": [
                    "profile"
                ],
                "summary": "Change user photo",
                "description": "Upload or update user photo as base64 string",
                "operationId": "changeUserPhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "example": "base64imagestring"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User photo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Photo updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "photo_id"
                                                },
                                                "images": {
                                                    "type": "string",
                                                    "example": "base64imagestring"
                                                },
                                                "created_by": {
                                                    "type": "string",
                                                    "example": "userA"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025/07/03 01:46:43"
                                                },
                                                "updated_by": {
                                                    "type": "string",
                                                    "example": "userA"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025/07/03 01:46:43"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Image is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}